Integrate `build_bus_regions` into `base_network`
Fixes a problem with the `build_bus_regions` writing to base.nc
without declaring base.nc as an output.
