+++ rules/build_electricity.smk
@@ -86,7 +86,9 @@ rule base_network:
         offshore_shapes=resources("offshore_shapes.geojson"),
         europe_shape=resources("europe_shape.geojson"),
     output:
-        resources("networks/base.nc"),
+        base_network=resources("networks/base.nc"),
+        regions_onshore=resources("regions_onshore.geojson"),
+        regions_offshore=resources("regions_offshore.geojson"),
     log:
         logs("base_network.log"),
     benchmark:
@@ -125,27 +127,6 @@ rule build_shapes:
         "../envs/environment.yaml"
     script:
         "../scripts/build_shapes.py"
-
-
-rule build_bus_regions:
-    params:
-        countries=config_provider("countries"),
-    input:
-        country_shapes=resources("country_shapes.geojson"),
-        offshore_shapes=resources("offshore_shapes.geojson"),
-        base_network=resources("networks/base.nc"),
-    output:
-        regions_onshore=resources("regions_onshore.geojson"),
-        regions_offshore=resources("regions_offshore.geojson"),
-    log:
-        logs("build_bus_regions.log"),
-    threads: 1
-    resources:
-        mem_mb=1000,
-    conda:
-        "../envs/environment.yaml"
-    script:
-        "../scripts/build_bus_regions.py"
 
 
 if config["enable"].get("build_cutout", False):
